OverClocked
This was a project that ran from November 2011 to January 2012. It is a
2D side scrolling game that focuses on action and insanity. The game was
designed to be impossible to beat without aquirring "power ups" from
playing minigames(H.A.C.K Training). The "power ups" are very similar to
hacks you would see in games like Maplestory. The idea sparked from our
realization that abusing and breaking down a game by hacking is FUN.