Quandary
Quandary was our first Android Game and mimics the famous 15 Puzzle. We
could technically expand the grid to an infinite size but we would
require more tile pictures and space...We implemented an algorithm to
check if the board was valid from
http://www.cs.bham.ac.uk/~mdr/teaching/modules04/java2/TilesSolvability.html
and the splash intro music is So Sick by Neyo.
This game was built to run on Android 2.3 and higher and on a screen
size 480x800.